Revert "Refactor RuleViolation back into a single class"
This reverts commit d7120c77c38760273a0e0f7cca56bc93f911258f.
